# docs/openapiSpec.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fashivo - Fashion E-commerce Website API",
    "version": "1.0.0",
    "description": "API for Fashivo fashion e-commerce website",
    "contact": {
      "name": "Huỳnh Thanh Nhuận, Thạch Nguyễn Khang",
      "email": "khangb2205987@student.ctu.edu.vn"
    }
  },
  "servers": [
    {
      "url": "https://api.fashion.drstone.id.vn",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/product/{id}": {
      "get": {
        "summary": "Get product by ID",
        "description": "Retrieves detailed information about a single product using its unique ID.",
        "operationId": "getProductById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/productIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Product data retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "product": {
                          "$ref": "#/components/schemas/Product"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Product"
        ]
      },
      "put": {
        "summary": "Update an existing product",
        "description": "Updates one or more fields of an existing product, including an optional new image upload.",
        "operationId": "updateProduct",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/productIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              },
              "encoding": {
                "imageFile": {
                  "contentType": "image/png, image/jpeg, image/gif"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Product updated successfully."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "product": {
                          "$ref": "#/components/schemas/Product"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Product"
        ]
      },
      "delete": {
        "summary": "Delete a product by ID",
        "description": "Deletes a specific product from the inventory using its unique ID.",
        "operationId": "deleteProduct",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/productIdParam"
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully. (No content returned)"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Product"
        ]
      }
    },
...
    "/api/v1/carts": {
      "get": {
        "summary": "Get all carts",
        "description": "Retrieve a list of all shopping carts with pagination and optional user ID filter.",
        "operationId": "getAllCarts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "description": "Filter carts by user ID."
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of carts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "carts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CartWithItems"
                          }
                        },
                        "metadata": {
                          "$ref": "#/components/schemas/PaginationMetadata"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Cart"
        ]
      },
      "post": {
        "summary": "Create or update a shopping cart",
        "description": "Adds items to a user's cart. If the cart doesn't exist, it creates one. If the product already exists in the cart, it updates the quantity. This operation expects multipart/form-data where 'items' is a JSON string.",
        "operationId": "createCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "format": "int32",
                    "description": "ID of the user for whom the cart is being created/updated.",
                    "minimum": 1
                  },
                  "items": {
                    "type": "string",
                    "description": "JSON string representing an array of cart items. Example: `[{\"product_id\": 1, \"quantity\": 2}, {\"product_id\": 3, \"quantity\": 1}]`",
                    "example": "[{\"product_id\": 1, \"quantity\": 2}, {\"product_id\": 3, \"quantity\": 1}]"
                  }
                },
                "required": [
                  "user_id",
                  "items"
                ]
              },
              "encoding": {
                "items": {
                  "contentType": "application/json"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cart created or updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cart": {
                          "$ref": "#/components/schemas/CartWithItems"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "summary": "Delete all shopping carts",
        "description": "Deletes all shopping carts and their items. (Use with caution)",
        "operationId": "deleteAllCarts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "All carts deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/api/v1/carts/user/{userId}": {
      "get": {
        "summary": "Get a user's shopping cart",
        "description": "Retrieve a specific user's shopping cart, including all items within it.",
        "operationId": "getCartByUserId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/userIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User's cart data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cart": {
                          "$ref": "#/components/schemas/CartWithItems"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/api/v1/carts/{id}": {
      "get": {
        "summary": "Get a shopping cart by ID",
        "description": "Retrieve a specific shopping cart by its ID, including all items within it.",
        "operationId": "getCartById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/cartIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cart": {
                          "$ref": "#/components/schemas/CartWithItems"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "summary": "Delete a shopping cart by ID",
        "description": "Deletes a shopping cart and all its associated items.",
        "operationId": "deleteCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/cartIdParam"
          }
        ],
        "responses": {
          "204": {
            "description": "Cart deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "type": {
            "type": "string",
            "example": "Electronics"
          },
          "name": {
            "type": "string",
            "example": "Smartphone"
          },
          "description": {
            "type": "string",
            "example": "Latest model smartphone",
            "nullable": true,
            "maxLength": 1000
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 699.99
          },
          "stock": {
            "type": "integer",
            "format": "int32",
            "example": 100
          },
          "available": {
            "type": "boolean",
            "example": true,
            "description": "True if stock > 0 or explicitly set"
          },
          "image_url": {
            "type": "string",
            "example": "/public/img/products/product-12345.jpg"
          },
          "category_id": {
            "type": "integer",
            "format": "int32",
            "description": "ID of product catalog which products belong to",
            "example": 1,
            "nullable": true
          },
          "category_name": {
            "type": "string",
            "readOnly": true,
            "description": "Description of the product category (read-only, returned from join)",
            "example": "High-performance computing devices."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "price",
          "stock",
          "available",
          "image_url",
          "created_at",
          "updated_at"
        ]
      },
...
    },
    "parameters": {
      "productIdParam": {
        "name": "id",
        "description": "Product ID",
        "required": true,
        "in": "path",
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
...
      "pageParam": {
        "name": "page",
        "description": "Page number of records",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        },
        "required": false
      },
...
      "userIdParam": {
        "name": "userId",
        "description": "User ID",
        "required": true,
        "in": "path",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "description": "Authenticated but no permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "NoPermission": {
                "summary": "Permission denied",
                "value": {
                  "status": "error",
                  "message": "You do not have permission to perform this action.",
                  "statusCode": 403
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "UserNotFound": {
                "summary": "User ID not found",
                "value": {
                  "status": "error",
                  "message": "User with ID 123 was not found.",
                  "statusCode": 404
                }
              },
              "ProductNotFound": {
                "summary": "Product not found",
                "value": {
                  "status": "error",
                  "message": "Product with ID 5 does not exist.",
                  "statusCode": 404
                }
              },
              "CategoryNotFound": {
                "summary": "Category not found",
                "value": {
                  "status": "error",
                  "message": "No category found with ID 2.",
                  "statusCode": 404
                }
              },
              "CartNotFound": {
                "summary": "Cart for user not found",
                "value": {
                  "status": "error",
                  "message": "Cart for user ID 7 not found.",
                  "statusCode": 404
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from login (e.g., `Bearer <token>`)"
      }
    }
  },
...
}